app.js 파일로 로직 분리
